tools/sync-kb-to-qdrant/pyproject.toml
[project]
name = "sync-data"
version = "0.1.0"
description = "Knowledge Base Data Sync to Qdrant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "jieba>=0.42.0",
]

[project.scripts]
sync-kb = "sync_data.main:main"
